requirements.txt
gradio
langchain
openai
python-dotenv